refactor: extract TrackHeart component
